# Improve hashing in dd.shuffle