Previously, python's `hash` function was used to hash indices for partitioning in shuffle. This could lead to non-ideal behavior, as python's hash function could have low entropy (on my machine `hash((a,
b)) % 5 == 1` for any integers `a` and `b`). This also was slow-ish for large chunks, as it doesn't take advantage of numpy/pandas vectorized operations.

This implements a custom elementwise hashing operation for use in shuffle partitioning. Each pandas dtype has a small function to map a series of that type deterministically to an integer array (the hashes),
which are then combined iteratively with a simple `new_hash = 3 * hash1 + hash2`. This is both faster (as it takes advantage of vectorization for non object columns), and seems to partition more evenly in many cases.

This also fixes a bug where empty partitions would lack the requisite metadata (column names or dtypes), which caused problems with subsequent operations.