@mrocklin 2016-04-13
Nice.  Do you happen to have benchmark numbers for an on-disk groupby?