@jcrist 2016-04-14
Here's one for just the shuffle. Admittedly the bottleneck isn't the hash, but you can still see the performance difference.

**Master:**

```python
In [1]: import dask.dataframe as dd

In [2]: df = dd.demo.make_timeseries('2000', '2010', {'value': float, 'id': int}, freq='10s', partition_freq='365D', seed=1)

In [3]: len(df)
Out[3]: 31536000

In [4]: s = dd.shuffle.shuffle(df, df.id)

In [5]: %time _ = s.compute()
CPU times: user 44.4 s, sys: 7.23 s, total: 51.6 s
Wall time: 43.2 s
```

**This PR:**

```python
In [1]: import dask.dataframe as dd

In [2]: df = dd.demo.make_timeseries('2000', '2010', {'value': float, 'id': int}, freq='10s', partition_freq='365D', seed=1)

In [3]: len(df)
Out[3]: 31536000

In [4]: s = dd.shuffle.shuffle(df, df.id)

In [5]: %time _ = s.compute()
CPU times: user 30.2 s, sys: 2.84 s, total: 33.1 s
Wall time: 27.1 s
```