@mrocklin 2016-04-14
This all seems sensible to me.  +1